refactor(55-04): split DetailsDialog.tsx into details/ sub-components
- Create details/DetailsPrimitives.tsx: CopyableValue, DetailRow, formatDateWithTime
- Create details/VersionHistory.tsx: preserves void folderKey lint suppression (Pitfall 7)
- Create details/FileDetails.tsx: file metadata display with version history
- Create details/FolderDetails.tsx: folder IPNS/key/timestamp display
- Slim DetailsDialog.tsx to container: retains both cross-guarded useEffects (Pitfall 4)
- CSS import stays in container only; barrel (index.ts) untouched
- Container props/signature byte-identical; 63 web tests pass; tsc --noEmit clean
