+++ apps/web/src/components/file-browser/details/FolderDetails.tsx
@@ -1,0 +1,93 @@
+import type { FolderEntry } from '@cipherbox/core';
+import { formatDate } from '../../../utils/format';
+import { CopyableValue, DetailRow } from './DetailsPrimitives';
+
+/**
+ * Folder details content.
+ */
+export function FolderDetails({
+  item,
+  metadataCid,
+  metadataLoading,
+  sequenceNumber,
+  childCount,
+}: {
+  item: FolderEntry;
+  metadataCid: string | null;
+  metadataLoading: boolean;
+  sequenceNumber: bigint | null;
+  childCount: number | null;
+}) {
+  return (
+    <div className="details-rows">
+      <DetailRow label="Name">
+        <span className="details-value">{item.name}</span>
+      </DetailRow>
+
+      <DetailRow label="Type">
+        <span className="details-type-badge details-type-badge--folder">[DIR]</span>
+      </DetailRow>
+
+      <DetailRow label="Contents">
+        {childCount !== null ? (
+          <span className="details-value">
+            {childCount} {childCount === 1 ? 'item' : 'items'}
+          </span>
+        ) : (
+          <span className="details-value details-value--dim">unknown</span>
+        )}
+      </DetailRow>
+
+      {/* IPNS section */}
+      <div className="details-section-header">{'// ipns'}</div>
+
+      <DetailRow label="IPNS Name">
+        <CopyableValue value={item.ipnsName} />
+      </DetailRow>
+
+      <DetailRow label="Metadata CID">
+        {metadataLoading ? (
+          <span className="details-loading">resolving...</span>
+        ) : metadataCid ? (
+          <CopyableValue value={metadataCid} />
+        ) : (
+          <span className="details-value details-value--dim">unavailable</span>
+        )}
+      </DetailRow>
+
+      <DetailRow label="Sequence Number">
+        <span className="details-value">
+          {sequenceNumber !== null ? sequenceNumber.toString() : '—'}
+        </span>
+      </DetailRow>
+
+      {/* Crypto section */}
+      <div className="details-section-header">{'// encryption'}</div>
+
+      <DetailRow label="Folder Key">
+        <span className="details-value details-value--redacted">
+          {item.folderKeyEncrypted.slice(0, 16)}...{item.folderKeyEncrypted.slice(-8)}{' '}
+          (ECIES-wrapped)
+        </span>
+      </DetailRow>
+
+      <DetailRow label="IPNS Private Key">
+        <span className="details-value details-value--redacted">
+          {item.ipnsPrivateKeyEncrypted.slice(0, 16)}...{item.ipnsPrivateKeyEncrypted.slice(-8)}{' '}
+          (ECIES-wrapped)
+        </span>
+      </DetailRow>
+
+      {/* Timestamps */}
+      <div className="details-section-header">{'// timestamps'}</div>
+
+      <DetailRow label="Created">
+        <span className="details-value">{formatDate(item.createdAt)}</span>
+      </DetailRow>
+
+      <DetailRow label="Modified">
+        <span className="details-value">{formatDate(item.modifiedAt)}</span>
+      </DetailRow>
+    </div>
+  );
+}
